+++ src/main/java/com/korbyte/alphavantage/core/monthly_adjusted/MonthlyAdjustedPositionData.java
@@ -12,7 +12,7 @@ public class MonthlyAdjustedPositionData extends BasePositionData {
   private Float adjustedClose;
 
   @JsonProperty(required = true, value = "6. volume")
-  private Double volume;
+  private Integer volume;
 
   @JsonProperty(required = true, value = "7. dividend amount")
   private Float dividendAmount;
